Working on embedding the ScheduleEvent use case within the View Calendar use case.
